requirements.txt
# Core dependencies
vllm>=0.6.0
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Code analysis
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-typescript>=0.21.0
pygments>=2.17.0

# Vector database
qdrant-client>=1.8.0
sentence-transformers>=2.5.0

# Database
asyncpg>=0.29.0
redis>=5.0.0
hiredis>=2.3.0

# Monitoring
prometheus-client>=0.20.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
httpx>=0.27.0
aiofiles>=23.2.0
tenacity>=8.2.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.3.0
mypy>=1.8.0
